moves method TestnetInPlaceCommand to testnet file
